+++ tests/files/comments-function-args/output.lua
@@ -1,0 +1,27 @@
+foo( -- comment
+	baz
+)
+
+foo(
+	baz, -- comment
+	bar
+)
+
+foo(
+	baz,
+	-- comment
+	bar
+)
+
+foo(
+	baz,
+	bar -- comment
+)
+
+foo(baz, bar) -- comment
+
+foo(
+	-- comment
+	baz,
+	bar
+)