+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Fixed function calls being incorrectly expanded due to a comment within the arguments.
+We will now only check for leading/trailing comments for argument expressions to see if we need to keep it expanded or not.
 
 ## [0.4.0] - 2021-02-05
 ### Added